requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
commonmark==0.9.1
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
ipapi==1.0.4
outcome==1.2.0
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PySocks==1.7.1
PyYAML==6.0
requests==2.28.1
rich==12.6.0
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
wsproto==1.2.0
